backend/requirements.txt
Flask
Flask-Cors
numpy
qiskit>=0.45.0
qiskit-aer>=0.13.0
qiskit-algorithms>=0.3.0
requests
python-dotenv
scipy
gunicorn
